# artifacts/api-server/src/routes/candidates/index.ts
import { Router, type IRouter } from "express";
import { db } from "@workspace/db";
import { candidatesTable, candidatePipelineTable } from "@workspace/db/schema";
import { eq, and, isNull, ilike, or } from "drizzle-orm";
import { requireAuth } from "../../middleware/auth";

const router: IRouter = Router();

router.get("/candidates", requireAuth, async (req, res): Promise<void> => {
  const { search, source, location, limit = "50", offset = "0" } = req.query;

  const conditions = [isNull(candidatesTable.deletedAt)];
  if (source && typeof source === "string") conditions.push(eq(candidatesTable.source, source as any));
  if (location && typeof location === "string") conditions.push(ilike(candidatesTable.location, `%${location}%`));
  if (search && typeof search === "string") {
    conditions.push(
      or(
        ilike(candidatesTable.name, `%${search}%`),
        ilike(candidatesTable.email, `%${search}%`),
        ilike(candidatesTable.currentRole, `%${search}%`),
      )!
    );
  }

  const rows = await db
    .select()
    .from(candidatesTable)
    .where(and(...conditions))
    .limit(Number(limit))
    .offset(Number(offset));

  res.json(rows);
});

router.post("/candidates", requireAuth, async (req, res): Promise<void> => {
  const { name, email, phone, currentRole, experienceYears, location, skills, source, currentCtc, expectedCtc, noticeDays, summary } = req.body;

  if (!name || !email) {
    res.status(400).json({ error: "Name and email are required" });
    return;
  }

  const initials = name.split(" ").map((w: string) => w[0]).join("").slice(0, 2).toUpperCase();

  const [candidate] = await db
    .insert(candidatesTable)
    .values({
      name, email, phone, currentRole,
      experienceYears: experienceYears ? Number(experienceYears) : undefined,
      location, skills, source: source || "portal",
      currentCtc: currentCtc ? Number(currentCtc) : undefined,
      expectedCtc: expectedCtc ? Number(expectedCtc) : undefined,
      noticeDays: noticeDays ? Number(noticeDays) : undefined,
      summary, initials,
    })
    .returning();

  res.status(201).json(candidate);
});

router.get("/candidates/:id", requireAuth, async (req, res): Promise<void> => {
  const id = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;

  const [candidate] = await db
    .select()
    .from(candidatesTable)
    .where(and(eq(candidatesTable.id, id), isNull(candidatesTable.deletedAt)));

  if (!candidate) {
    res.status(404).json({ error: "Candidate not found" });
    return;
  }
  res.json(candidate);
});

router.patch("/candidates/:id", requireAuth, async (req, res): Promise<void> => {
  const id = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const { name, email, phone, currentRole, experienceYears, location, skills, source, currentCtc, expectedCtc, noticeDays, summary } = req.body;

  const [updated] = await db
    .update(candidatesTable)
    .set({ name, email, phone, currentRole, experienceYears, location, skills, source, currentCtc, expectedCtc, noticeDays, summary, updatedAt: new Date() })
    .where(and(eq(candidatesTable.id, id), isNull(candidatesTable.deletedAt)))
    .returning();

  if (!updated) {
    res.status(404).json({ error: "Candidate not found" });
    return;
  }
  res.json(updated);
});

router.delete("/candidates/:id", requireAuth, async (req, res): Promise<void> => {
  const id = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  await db.update(candidatesTable).set({ deletedAt: new Date() }).where(eq(candidatesTable.id, id));
  res.sendStatus(204);
});

router.post("/candidates/:id/apply/:requirementId", requireAuth, async (req, res): Promise<void> => {
  const candidateId = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const requirementId = Array.isArray(req.params.requirementId) ? req.params.requirementId[0] : req.params.requirementId;

  const [entry] = await db
    .insert(candidatePipelineTable)
    .values({
      candidateId,
      requirementId,
      stage: "sourced",
      assignedRecruiterId: req.employee?.employeeId,
    })
    .returning();

  res.status(201).json(entry);
});

export default router;