# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: 99Placement Recruitment OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Logged out

  # ── COMPANIES ───────────────────────────────────────────────────────────────
...
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List candidates
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: stage
          in: query
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Candidates list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Create a candidate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCandidateRequest"
      responses:
        "201":
          description: Created candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get candidate with full profile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update candidate profile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCandidateRequest"
      responses:
        "200":
          description: Updated candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
    delete:
      operationId: deleteCandidate
      tags: [candidates]
      summary: Delete candidate (soft)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /candidates/{id}/apply/{requirementId}:
    post:
      operationId: applyCandidate
      tags: [candidates]
      summary: Add candidate to requirement pipeline
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: requirementId
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Pipeline entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineEntry"

  # ── PIPELINE ─────────────────────────────────────────────────────────────────
...
  /assessment-questions:
    get:
      operationId: listAssessmentQuestions
      tags: [assessments]
      summary: List questions from question bank
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssessmentQuestion"
    post:
      operationId: createAssessmentQuestion
      tags: [assessments]
      summary: Add question to bank
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionRequest"
      responses:
        "201":
          description: Created question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentQuestion"
  /assessments:
    post:
      operationId: createAssessmentTest
      tags: [assessments]
      summary: Create a test for a candidate pipeline entry
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssessmentTestRequest"
      responses:
        "201":
          description: Created test with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentTestWithQuestions"
  /assessments/{id}:
    get:
      operationId: getAssessmentTest
      tags: [assessments]
      summary: Get test and results
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Assessment test
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentTestWithQuestions"
  /assessments/{id}/submit:
    post:
      operationId: submitAssessment
      tags: [assessments]
      summary: Submit answers and get score
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAssessmentRequest"
      responses:
        "200":
          description: Scored assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentTestWithQuestions"

  # ── CLIENT INTERVIEWS ────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # AUTH ──────────────────────────────────────────────────────────────────────
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        employee:
          $ref: "#/components/schemas/Employee"
      required: [token, employee]
    Employee:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, recruiter, hr, client]
        companyId:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, isActive, createdAt]

    # COMPANIES ─────────────────────────────────────────────────────────────────
...
    Candidate:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        initials:
          type: string
          nullable: true
        email:
          type: string
        phone:
          type: string
          nullable: true
        currentRole:
          type: string
          nullable: true
        experienceYears:
          type: integer
          nullable: true
        location:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
          nullable: true
        source:
          type: string
          enum: [referral, portal, social, internal, direct]
        currentCtc:
          type: integer
          nullable: true
        expectedCtc:
          type: integer
          nullable: true
        noticeDays:
          type: integer
          nullable: true
        summary:
          type: string
          nullable: true
        resumeUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, source, createdAt]
    CreateCandidateRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        currentRole:
          type: string
        experienceYears:
          type: integer
        location:
          type: string
        skills:
          type: array
          items:
            type: string
        source:
          type: string
          enum: [referral, portal, social, internal, direct]
        currentCtc:
          type: integer
        expectedCtc:
          type: integer
        noticeDays:
          type: integer
        summary:
          type: string
      required: [name, email]

    # PIPELINE ──────────────────────────────────────────────────────────────────
    PipelineEntry:
      type: object
      properties:
        id:
          type: string
        candidateId:
          type: string
        requirementId:
          type: string
        stage:
          type: string
          enum: [sourced, screened, assessed, shortlisted, client_interview, offer, joining, post_joining, rejected, dropped]
        assignedRecruiterId:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        stageUpdatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, candidateId, requirementId, stage, stageUpdatedAt, createdAt]
...
    AssessmentQuestion:
      type: object
      properties:
        id:
          type: string
        category:
          type: string
          enum: [aptitude, mathematics, english, logical_reasoning, computer_knowledge, general_knowledge, current_affairs, technical]
        questionText:
          type: string
        options:
          type: array
          items:
            type: string
        correctOption:
          type: integer
        difficulty:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, category, questionText, options, correctOption, difficulty, isActive, createdAt]
    CreateQuestionRequest:
      type: object
      properties:
        category:
          type: string
          enum: [aptitude, mathematics, english, logical_reasoning, computer_knowledge, general_knowledge, current_affairs, technical]
        questionText:
          type: string
        options:
          type: array
          items:
            type: string
          minItems: 2
        correctOption:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard]
      required: [category, questionText, options, correctOption]
    AssessmentTestWithQuestions:
      type: object
      properties:
        id:
          type: string
        pipelineId:
          type: string
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        totalQuestions:
          type: integer
        correctAnswers:
          type: integer
        totalScore:
          type: integer
        maxScore:
          type: integer
        percentage:
          type: integer
        verdict:
          type: string
          nullable: true
        questions:
          type: array
          items:
            $ref: "#/components/schemas/AssessmentQuestion"
        createdAt:
          type: string
          format: date-time
      required: [id, pipelineId, totalQuestions, correctAnswers, totalScore, maxScore, percentage, questions, createdAt]
    CreateAssessmentTestRequest:
      type: object
      properties:
        pipelineId:
          type: string
        categories:
          type: array
          items:
            type: string
          description: Which categories to include
        questionsPerCategory:
          type: integer
          default: 5
      required: [pipelineId]
    SubmitAssessmentRequest:
      type: object
      properties:
        answers:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: string
              selectedOption:
                type: integer
            required: [questionId, selectedOption]
      required: [answers]

    # CLIENT INTERVIEWS ──────────────────────────────────────────────────────────
...